requirements.txt
apiclient==1.0.4
appdirs==1.4.4
autopep8==2.0.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
coloredlogs==14.3
decorator==4.4.2
docopt==0.6.2
fs==2.4.16
google-api-core==2.11.0
google-api-python-client==2.74.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.58.0
httplib2==0.21.0
humanfriendly==10.0
idna==3.4
imageio==2.24.0
imageio-ffmpeg==0.4.8
instaloader==4.9.5
moviepy==1.0.3
numpy==1.24.1
oauthlib==3.2.2
Pillow==9.4.0
proglog==0.1.10
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pyparsing==3.0.9
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tenacity==6.3.1
tomli==2.0.1
tqdm==4.64.1
uritemplate==4.1.1
urllib3==1.26.14
verboselogs==1.7
